refactor: HttpService
