+++ apps/app/src/lib/services/HttpServiceWebLive.ts
@@ -1,0 +1,25 @@
+import { HttpService } from '$lib/services/HttpService';
+import { HttpClient, HttpClientRequest, HttpClientResponse } from '@effect/platform';
+import { WhisperingError } from '@repo/shared';
+import { Effect, Layer } from 'effect';
+
+export const HttpServiceWebLive = Layer.succeed(
+	HttpService,
+	HttpService.of({
+		post: ({ formData, url, schema }) =>
+			HttpClientRequest.post(url).pipe(
+				HttpClientRequest.formDataBody(formData),
+				HttpClient.fetch,
+				Effect.andThen(HttpClientResponse.schemaBodyJson(schema)),
+				Effect.scoped,
+				Effect.mapError(
+					(error) =>
+						new WhisperingError({
+							title: 'Error transcribing audio',
+							description: error.message,
+							error,
+						}),
+				),
+			),
+	}),
+);
